Trace average FPS a) from start-to-finish of each song and b) every five seconds during each song.
This also includes a simpler FPS calculation than was previously there. This code should be mined for that logic, as it is a cleaner implementation for use even when not collecting average FPS stats or, for example, when only collecting them for each song as a whole but without the every-five-seconds variation.